// src/api/generated/models/companyProfileResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { UserResource } from './userResource';
import type { CompanyProfileResourceBioPt } from './companyProfileResourceBioPt';
import type { CompanyProfileResourceBioEn } from './companyProfileResourceBioEn';
import type { CompanyProfileResourceTranslationStatusLabel } from './companyProfileResourceTranslationStatusLabel';
import type { CompanyProfileResourceOperationalSegmentLabel } from './companyProfileResourceOperationalSegmentLabel';
import type { AddressResource } from './addressResource';
import type { CompanyProfileResourceCreatedAt } from './companyProfileResourceCreatedAt';
import type { CompanyProfileResourceUpdatedAt } from './companyProfileResourceUpdatedAt';

export interface CompanyProfileResource {
  id: string;
  user_id: string;
  user?: UserResource;
  name: string;
  bio: string;
  bio_pt: CompanyProfileResourceBioPt;
  bio_en: CompanyProfileResourceBioEn;
  translation_status: string;
  translation_status_label: CompanyProfileResourceTranslationStatusLabel;
  cnpj: string;
  phone: string;
  fouding_date: string;
  operational_segment: string;
  operational_segment_label: CompanyProfileResourceOperationalSegmentLabel;
  score: number;
  address?: AddressResource;
  is_blocked: boolean;
  created_at: CompanyProfileResourceCreatedAt;
  updated_at: CompanyProfileResourceUpdatedAt;
}